Update (search bar now generally works with console.log, already outsourced to components)
